requirements.txt
osiris-sdk
requests
python-dateutil